update. leo y extraigo valores enteros del comandos pero no puedo aplicarlos al sistema

+++ Core/Src/modulos/include/mmap.h
@@ -21,7 +21,7 @@
 // El contador de datos almacenados se guardara en buffer2
 
 // Guardo los primeros 10 registros para usos en configuraciones en el futuro
-#define MMAP_OFFSET_CONFIG                                  (10)
+#define MMAP_OFFSET_CONFIG                                  (uint16_t) 10
 
 // De 0 A 47   // Registros especiales, No Data
